pyproject.toml
[tool.poetry]
name = "habbitstracker"
version = "0.1.0"
description = ""
authors = ["Tsurkan <mikh.tsurkan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
requests = "^2.32.2"
celery = "^5.4.0"
redis = "^5.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
